updated pset5 tests
