+++ checks/cs50/2017/x/speller/basic
@@ -1,0 +1,8 @@
+brown
+dog
+fox
+jumps
+lazy
+over
+quick
+the
