+++ checks/cs50/2017/x/speller/basic.out
@@ -1,0 +1,6 @@
+MISSPELLED WORDS
+
+
+WORDS MISSPELLED:     0
+WORDS IN DICTIONARY:  8
+WORDS IN TEXT:        9
