+++ checks/cs50/2017/x/speller/basic.txt
@@ -1,0 +1,1 @@
+The quick brown fox jumps over the lazy dog.
