+++ checks/cs50/2017/x/speller/foo
@@ -1,0 +1,1 @@
+foo
