+++ checks/cs50/2017/x/speller/foo.txt
@@ -1,0 +1,8 @@
+foo
+foO
+fOo
+Foo
+fOO
+FoO
+FOo
+FOO
