+++ checks/cs50/2017/x/speller/foo_foos.out
@@ -1,0 +1,7 @@
+MISSPELLED WORDS
+
+foo's
+
+WORDS MISSPELLED:     1
+WORDS IN DICTIONARY:  1
+WORDS IN TEXT:        1
