+++ checks/cs50/2017/x/speller/foos
@@ -1,0 +1,1 @@
+foo's
