+++ checks/cs50/2017/x/speller/foos.txt
@@ -1,0 +1,1 @@
+foo's
