+++ checks/cs50/2017/x/speller/foos_foo.out
@@ -1,0 +1,14 @@
+MISSPELLED WORDS
+
+foo
+foO
+fOo
+Foo
+fOO
+FoO
+FOo
+FOO
+
+WORDS MISSPELLED:     8
+WORDS IN DICTIONARY:  1
+WORDS IN TEXT:        8
