+++ checks/cs50/2017/x/speller/max_length
@@ -1,0 +1,1 @@
+pneumonoultramicroscopicsilicovolcanoconiosis
