+++ checks/cs50/2017/x/speller/max_length.out
@@ -1,0 +1,6 @@
+MISSPELLED WORDS
+
+
+WORDS MISSPELLED:     0
+WORDS IN DICTIONARY:  1
+WORDS IN TEXT:        1
