+++ checks/cs50/2017/x/speller/max_length.txt
@@ -1,0 +1,1 @@
+pneumonoultramicroscopicsilicovolcanoconiosis
