+++ checks/cs50/2017/x/speller/min_length.txt
@@ -1,0 +1,1 @@
+a
