+++ checks/cs50/2017/x/speller/substring
@@ -1,0 +1,2 @@
+cat
+caterpillar
