+++ checks/cs50/2017/x/speller/substring.out
@@ -1,0 +1,10 @@
+MISSPELLED WORDS
+
+ca
+cats
+caterpill
+caterpillars
+
+WORDS MISSPELLED:     4
+WORDS IN DICTIONARY:  2
+WORDS IN TEXT:        6
