+++ checks/cs50/2017/x/speller/substring.txt
@@ -1,0 +1,1 @@
+ca cat cats caterpill caterpillar caterpillars
